+++ src/com/podalv/search/server/api/PatientComparator.java
@@ -23,9 +23,9 @@ public class PatientComparator {
 
     final int maxPatientCnt = Integer.parseInt(args[0]);
 
-    final LinkedList<Integer> pids = atlas1.getPatientIds(maxPatientCnt);
+    final LinkedList<Long> pids = atlas1.getPatientIds(maxPatientCnt);
 
-    for (final Integer pid : pids) {
+    for (final Long pid : pids) {
       if (atlas1.containsPatient(pid) && atlas2.containsPatient(pid)) {
         final DumpResponse p1 = atlas1.getPatientDumpResponse(pid);
         final DumpResponse p2 = atlas2.getPatientDumpResponse(pid);
